Today we start to complete the logic of the frmCreateTeam starting with the "CreatePerson".
Implemented a method to validate the data coming.
Implemented a clausule in the IDataConnection Interface to stabilish the obrigatority of the implementation of the CreatePerson method in the classes that asign to it

We correct the class PersonModel by adding the field "Id". It was missing in that class.

The method CreatePerson is already implemented in the class SqlConnector. We've made some tests to make sure that method actually saves the data in the SQL DataBase. The method also been implemented in the TextConnector to asure that we could save the data in a text document as well. We also have added the extend methods needed in the "TextConnectorProcessor" to support the method to create a person entry in the PeopleFile.csv file.