The stored procedure required to create a new team and its team members have been created in the database
We added a new property in the "TeamModel" class, now the "Id" field is available.

We also have created a new clause in the IDataConnection interface, by now ever class that signs this interface need to implement the "CreateTeam" method.

We've created the CreateTeam method in the SQL connector as well as we've implemented this method in the TextConnector as well.

A new table has been created in the tournaments database, a relational table called "TeamMembers". This table takes the TeamId and the PersonId to bind these two tables, respectively.

In the TextConnectorProcessor class a lot of new methods has been implemented to help the main method (CreateTeam) that saves the Team and TeamMembers in a text file. This new methods are:

ConvertToTeamModels()
ConvertPeopleListToString()
SaveToTeamsFile()