Tournament Tracker Commit 18th of October In this commit, we've already set up the expected behavior for the following buttons of the "frmCreateTournament":
- New team
- Create Team
- Create Prize
- Add Team
- Remove Selected Team (at the "TournamentTeamListBox")
- Remove Selected Prize( at the "TournamentPrizesListBox")

To make the feature of Create a Team and Create Prize possible, with lower coupling between the forms, we've implemented two new interfaces:

"IPrizeRequester" , "ITeamRequester"

These two interfaces implement the following methods, respectively:
 void PrizeComplete(PrizeModel model);
void TeamComplete(TeamModel model);
With that, we have done the encapsulating of the functionality of these two forms, and make it available for everyone that implements this two Interfaces.

By now the frmCreateTournament can load the available teams (in both databases, SQL and text) as well as it can pick the selected team and make it an actual player for the tournament.
The frmCreateTournament can also create teams, prizes and put it directly in the actual tournament teams and prizes.
We just need wire up the last button "Create Tournament".