Tournament Tracker Commit  - 24th October of 2018
In this commit, we've started to implement the feature to create a new tournament. This feature is a little bit trickier cause' it involves a lot of related data such as the prizes offered in the tournament and the teams competing.
A new clause is now available in the "IDataConnection" interface to standardize how the different data connectors perform the same task as well as we've implemented the code for this functionality in the SQL Connector, we split the task into three others private methods to help us to make the "Create Tournament" method more readable.
We have not yet implemented the counterpart in the Text Connector.