Tournament Tracker Commit - 25th Of October
In this commit, I've implemented the methods necessary to make the "TextConnector" able to persist the data from a tournament in the text Database.
A bunch of new methods has equals implemented in the "TextConnectorProcessor" class to make the "TextConnector" able to perform this task. The implementation of these methods are necessary because of the nature of the data enrolled in the Tournament data such as the Teams competing, the Prizes signed to the winners and so on.
We still need to implement the functionality of the "Create the rounds" and use it to populate the List of List of matchup models.